fix: refactor httpRequest bump package
